pyproject.toml
[project]
name = "asciipic"
description = "Image to ASCII renderer using Alex Harri's technique"
version = "0.1.2"
authors = [
    { name = "Gareth Davidson", email = "gaz@bitplane.net" }
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "Pillow",
]

[project.optional-dependencies]
train = [
    "torch",
    "tensorboard",
]
dev = [
    "pre-commit",
    "pytest",
    "coverage",
    "pytest-cov",
    "build",
    "twine",
    "ruff",
]

[project.scripts]
asciipic = "asciipic.cli:main"

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true
